Move IconData to egui, and use in ViewportBuilder
